Moving files to src for a better autoloader of files using composer
