+++ src/controllers/LogoutController.php
